requirements.txt
requests
streamlit
plotly
pandas
psycopg2-binary
sqlalchemy
python-dateutil
beautifulsoup4
lxml
cloudscraper
